refactor: Moved use of TimeAgo and relevant time functions to new component.  Moved formatDate and formatTime to files under utils.
