+++ src/utils/formatDate.ts
@@ -1,0 +1,7 @@
+export default function formatDate(date: Date): string {
+  return new Intl.DateTimeFormat('en-US', {
+    month: 'short',
+    day: 'numeric',
+    year: 'numeric'
+  }).format(date);
+}
