+++ src/utils/formatTime.ts
@@ -1,0 +1,6 @@
+export default function formatTime(date: Date): string {
+  return new Intl.DateTimeFormat('en-US', {
+    hour: 'numeric',
+    minute: 'numeric'
+  }).format(date).replace(' ', '').toLowerCase();
+}
